+++ src/renderer/App.tsx
@@ -78,7 +78,7 @@ const WarcraftRecorder = () => {
   const [videoState, setVideoState] = useState<RendererVideo[]>([]);
 
   const stateManager = useRef<StateManager>(
-    StateManager.getInstance(setVideoState, appState, setAppState),
+    StateManager.getInstance(setVideoState),
   );
 
   // Used to allow for hot switching of video players when moving between POVs.
@@ -220,15 +220,6 @@ const WarcraftRecorder = () => {
     ipc.on('updateCloudStatus', updateCloudStatus);
     ipc.on('updateAvailable', onUpdateAvailable);
   }, []);
-
-  // Debugging why we needed this hurt. I think it's because when we call setAppState, it sets
-  // appState to undefined and reassigns it in this component. However that leaves the StateManager
-  // singleton with a reference pointing to undefined which breaks the frontend. So here we reapply
-  // the appState to the StateManager every time it updates. This is almost certainly massively
-  // overengineered but for now it works.
-  useEffect(() => {
-    stateManager.current.updateAppState(appState);
-  }, [appState]);
 
   return (
     <Box